+++ cast/admin.py
@@ -2,7 +2,7 @@ import logging
 
 from django.contrib import admin
 
-from .models import Blog, Post, File, Image, Video, Gallery
+from .models import Blog, Post, File, Image, Video, Audio, Gallery, ItunesArtWork
 
 logger = logging.getLogger(__name__)
 
@@ -37,12 +37,28 @@ class ImageModelAdmin(AdminUserMixin, admin.ModelAdmin):
 admin.site.register(Image, ImageModelAdmin)
 
 
-class FileAdmin(AdminUserMixin, admin.ModelAdmin):
+class ItunesArtWorkModelAdmin(AdminUserMixin, admin.ModelAdmin):
+    list_display = ("pk", "original")
+    fields = ("original",)
+
+
+admin.site.register(ItunesArtWork, ItunesArtWorkModelAdmin)
+
+
+class FileModelAdmin(AdminUserMixin, admin.ModelAdmin):
     list_display = ("original", "user")
     fields = ("user", "original")
 
 
-admin.site.register(File, FileAdmin)
+admin.site.register(File, FileModelAdmin)
+
+
+class AudioAdmin(AdminUserMixin, admin.ModelAdmin):
+    list_display = ("user", "title", "subtitle", "m4a", "mp3", "oga", "opus")
+    fields = ("user", "title", "subtitle", "m4a", "mp3", "oga", "opus")
+
+
+admin.site.register(Audio, AudioAdmin)
 
 
 class VideoModelAdmin(AdminUserMixin, admin.ModelAdmin):