+++ cast/templatetags/cast_extras.py
@@ -9,7 +9,7 @@ def image(context, pk):
     image = context["image"][pk]
     image_tag = (
         '<a href="{full}">'
-        '  <img class="blog-image" src="{src}" srcset="{srcset}" sizes="100vw"></img>'
+        '  <img class="cast-image" src="{src}" srcset="{srcset}" sizes="100vw"></img>'
         "</a>"
     ).format(srcset=image.get_srcset(), src=image.img_xs.url, full=image.img_full.url)
     return mark_safe(image_tag)
@@ -21,16 +21,22 @@ def video(context, pk):
     if video.poster:
         poster_url = video.poster.url
     else:
-        poster_url = "/static/images/Video-icon.svg"
+        poster_url = "/static/img/cast/Video-icon.svg"
     video_tag = (
-        '<video class="blog-video" preload="auto" controls poster="{poster}">'
+        '<video class="cast-video" preload="auto" controls poster="{poster}">'
         '  <source src="{src}" type="video/mp4">'
         "</video>"
     ).format(src=video.original.url, poster=poster_url)
     return mark_safe(video_tag)
 
 
-# blog_gallery tag
+@register.simple_tag(takes_context=True)
+def audio(context, pk):
+    audio_tag = f'<div id="audio_{pk}"></div>'
+    return mark_safe(audio_tag)
+
+
+# gallery tag
 
 
 def get_modal_trigger(gallery_key, image, prev_img, next_img):
@@ -38,7 +44,7 @@ def get_modal_trigger(gallery_key, image, prev_img, next_img):
     prev_id = "img-{}".format(prev_img.pk) if prev_img is not None else "false"
     next_id = "img-{}".format(next_img.pk) if next_img is not None else "false"
     thumbnail_tag = (
-        '<img id="img-{img_id}" class="gallery-thumbnail" src="{src}" '
+        '<img id="img-{img_id}" class="cast-gallery-thumbnail" src="{src}" '
         'srcset="{srcset}" data-prev="{prev}" data-next="{next}" '
         'data-full="{full}"></img>'
     ).format(
@@ -50,7 +56,7 @@ def get_modal_trigger(gallery_key, image, prev_img, next_img):
         full=image.img_full.url,
     )
     return """
-        <a src="#" class="gallery-modal" data-toggle="modal" data-target="#galleryModal{key}">
+        <a src="#" class="cast-gallery-modal" data-toggle="modal" data-target="#galleryModal{key}">
             {thumbnail_tag}
         </a>
     """.format(
@@ -61,7 +67,7 @@ def get_modal_trigger(gallery_key, image, prev_img, next_img):
 def get_image_thumb(image):
     srcset = image.get_srcset()
     thumbnail_tag = (
-        '<img class="gallery-thumbnail" src={src} ' 'srcset="{srcset}"</img>'
+        '<img class="cast-gallery-thumbnail" src={src} ' 'srcset="{srcset}"</img>'
     ).format(src=image.img_xs.url, srcset=srcset)
     return """
         <a href="{full}"">
@@ -78,17 +84,17 @@ def get_modal_tmpl():
         <!-- Modal -->
         <div class="modal fade" id="galleryModal{key}" tabindex="-1" role="dialog"
              aria-labelledby="galleryModalLabel" aria-hidden="true">
-            <div class="modal-dialog modal-lg gallery-lg" role="document">
-                <div class="modal-content gallery-content">
-                    <div class="modal-header gallery-header">
+            <div class="modal-dialog modal-lg cast-gallery-lg" role="document">
+                <div class="modal-content cast-gallery-content">
+                    <div class="modal-header cast-gallery-header">
                         <button type="button" class="close" data-dismiss="modal" aria-label="Close">
                             <span aria-hidden="true">&times;</span>
                             </button>
                         </div>
-                        <div class="modal-body gallery-body">
-                        <a href=""><img class="modal-image blog-image" src="" srcset="" sizes="100vw"></img></a>
+                        <div class="modal-body cast-gallery-body">
+                        <a href=""><img class="modal-image cast-image" src="" srcset="" sizes="100vw"></img></a>
                     </div>
-                    <div class="modal-footer gallery-footer">
+                    <div class="modal-footer cast-gallery-footer">
                         </div>
                     </div>
             </div>
@@ -99,7 +105,7 @@ def get_modal_tmpl():
 def gallery_with_javascript(gallery, post):
     image_thumbs = [
         "<!-- Button trigger modal -->",
-        '<div class="blog-gallery-container">',
+        '<div class="cast-gallery-container">',
     ]
 
     gallery_key = "{}_{}".format(post.pk, gallery.pk)