+++ cast/urls.py
@@ -1,10 +1,30 @@
-from django.conf.urls import url, include
+from django.conf.urls import url
+from django.urls import path, include
 
 from . import views
+from . import feeds
 
 app_name = "cast"
 urlpatterns = [
+    # API
     url(r"^api/", include("cast.api.urls", namespace="api")),
+    # Feeds
+    path(
+        "<slug:slug>/feed/rss.xml",
+        view=feeds.LatestEntriesFeed(),
+        name="latest_entries_feed",
+    ),
+    path(
+        "<slug:slug>/feed/podcast/<audio_format>/rss.xml",
+        view=feeds.RssPodcastFeed(),
+        name="podcast_feed_rss",
+    ),
+    path(
+        "<slug:slug>/feed/podcast/<audio_format>/atom.xml",
+        view=feeds.AtomPodcastFeed(),
+        name="podcast_feed_atom",
+    ),
+    # Regular django views
     url(
         regex=r"^(?P<slug>[^/]+)/add/$",
         view=views.PostCreateView.as_view(),
@@ -24,11 +44,6 @@ urlpatterns = [
         regex=r"^(?P<slug>[^/]+)/$",
         view=views.PostsListView.as_view(),
         name="post_list",
-    ),
-    url(
-        regex=r"^(?P<slug>[^/]+)/feed.xml$",
-        view=views.LatestEntriesFeed(),
-        name="post_feed",
     ),
     url(
         regex=r"^(?P<slug>[^/]+)_detail/$",