+++ cast/views.py
@@ -1,18 +1,12 @@
 import logging
 
+from django.utils import timezone
+from django.shortcuts import get_object_or_404
+from django.contrib.auth.mixins import LoginRequiredMixin
 from django.views.generic import ListView, DetailView, CreateView, UpdateView
 
-from django.contrib.syndication.views import Feed
-from django.contrib.auth.mixins import LoginRequiredMixin
-
-from django.urls import reverse
-from django.shortcuts import get_object_or_404
-
-
 from .forms import PostForm
-
 from .models import Blog, Post
-
 from .viewmixins import RenderPostMixin, AddRequestUserMixin, PostChangeMixin
 
 logger = logging.getLogger(__name__)
@@ -52,31 +46,6 @@ class PostsListView(RenderPostMixin, ListView):
         return context
 
 
-class LatestEntriesFeed(RenderPostMixin, Feed):
-    def get_object(self, request, *args, **kwargs):
-        self.object = get_object_or_404(Blog, slug=kwargs["slug"])
-
-    def title(self):
-        return self.object.title
-
-    def description(self):
-        return self.object.description
-
-    def link(self):
-        return reverse("cast:post_feed", kwargs={"slug": self.object.slug})
-
-    def items(self):
-        queryset = Post.published.filter(blog=self.object).order_by("-pub_date")
-        return queryset[:5]
-
-    def item_title(self, item):
-        return item.title
-
-    def item_description(self, item):
-        self.render_post(item, javascript=False)
-        return item.description
-
-
 class PostDetailView(RenderPostMixin, DetailView):
     model = Post
     template_name = "cast/post_detail.html"
@@ -106,6 +75,11 @@ class PostCreateView(
     template_name = "cast/post_edit.html"
     user_field_name = "author"
     success_msg = "Entry created!"
+
+    def get_initial(self):
+        initial = super().get_initial()
+        initial["visible_date"] = timezone.now()
+        return initial
 
     def form_valid(self, form):
         self.blog_slug = self.kwargs["slug"]
@@ -140,4 +114,5 @@ class PostUpdateView(
         context = super().get_context_data(**kwargs)
         # needed for back button
         context["blog_slug"] = self.kwargs["blog_slug"]
+        print("context post: ", context["object"].visible_date)
         return context