+++ HISTORY.rst
@@ -1,7 +1,44 @@
 .. :changelog:
 
 History
--------
+
+0.1.7 (2018-02-28)
+++++++++++++++++++
+
+* forgot linting
+
+0.1.6 (2018-02-28)
+++++++++++++++++++
+
+* Use filepond for media uploads (images video)
+* Improved portrait video support
+* Get api prefix programatically from schema
+* Fixed link to podcast in itunes (was feed, now it's post list)
+* Set visible date to now if it's not set
+* use load static instead of staticfiles (deprecated)
+* Fixed language displayed in itunes (you have to set it in base.py in settings)
+* Dont try to be fancy, just display a plain list of feed on top of post list site (and podcast feeds only if blog.is_podcast is True)
+
+0.1.5 (2018-11-21)
+++++++++++++++++++
+
+* basic feed support (rss/atom) for podcasts
+* travis now runs tests with ffprobe, too
+* documentation fixes from @SmartC2016 and @oryon-dominik
+
+0.1.4 (2018-11-18)
+++++++++++++++++++
+
+* Include css via cast_base.html
+* audio fixes
+
+0.1.3 (2018-11-17)
+++++++++++++++++++
+
+* Fixed css/static icons
+* Merged pull request from SmartC2016 to fix javascript block issue
+* Added some documentation
+
 0.1.2 (2018-11-08)
 ++++++++++++++++++